+++ resources/js/components/twc-ui/date-field.tsx
@@ -15,10 +15,10 @@ import {
   Text,
   type ValidationResult
 } from 'react-aria-components'
-import { useFormContext } from '@/components/twc-ui/form'
 import { useDateConversion } from '@/hooks/use-date-conversion'
 import { cn } from '@/lib/utils'
 import { FieldError, FormFieldError, fieldGroupVariants, Label } from './field'
+import { useFormContext } from './form'
 
 const BaseDateField = AriaDateField
 
@@ -115,7 +115,15 @@ const FormDateField = ({ value, onChange, ...props }: FormDateFieldProps) => {
   const error = form?.errors?.[props.name as string]
   const { parsedDate, handleChange } = useDateConversion(value, onChange)
 
-  return <DateField errorComponent={FormFieldError} errorMessage={error} value={parsedDate} onChange={handleChange} {...props} />
+  return (
+    <DateField
+      errorComponent={FormFieldError}
+      errorMessage={error}
+      value={parsedDate}
+      onChange={handleChange}
+      {...props}
+    />
+  )
 }
 
 export { DateField, FormDateField, DateSegment, DateInput, BaseDateField }