+++ resources/js/components/twc-ui/text-field.tsx
@@ -19,7 +19,7 @@ const Input = ({ className, ...props }: AriaInputProps) => {
     <AriaInput
       className={composeRenderProps(className, className =>
         cn(
-          'flex h-9 w-full rounded-sm border border-input bg-transparent px-3 py-1 font-medium shadow-none outline-0 transition-colors file:border-0 file:bg-transparent file:font-medium file:text-sm placeholder:text-muted-foreground',
+          'flex h-9 w-full rounded-sm border border-input bg-transparent px-3 py-1 font-medium text-sm shadow-none outline-0 transition-colors file:border-0 file:bg-transparent file:font-medium file:text-sm placeholder:text-muted-foreground',
           /* Disabled */
           'data-disabled:cursor-not-allowed data-disabled:opacity-50',
           /* Focused */
@@ -69,7 +69,7 @@ const TextField = ({
   return (
     <AriaTextField
       className={composeRenderProps(className, className =>
-        cn('group flex flex-col gap-1.5 text-sm', className)
+        cn('group flex flex-col gap-1.5', className)
       )}
       isInvalid={hasError}
       value={value ?? undefined}