+++ resources/js/pages/demos/icon/default.tsx
@@ -1,7 +1,7 @@
 import { Rocket02Icon } from '@hugeicons/core-free-icons'
+import { Moon } from 'lucide-react'
 import { DemoContainer } from '@/components/docs/DemoContainer'
 import { Button } from '@/components/twc-ui/button'
-import { Moon } from 'lucide-react'
 import { Icon } from '@/components/twc-ui/icon'
 
 export const Demo = () => {